Rename storage-db to storageDb
